+++ src/Auditable.php
@@ -71,7 +71,7 @@ trait Auditable
     /**
      * {@inheritdoc}
      */
-    public function audits(): MorphMany
+    public function histories(): MorphMany
     {
         return $this->morphMany(
             Config::get('audit.implementation', Models\Audit::class),